modified front-end for product list

+++ src/resources/views/product/list/index.blade.php
@@ -101,28 +101,30 @@
                                 </tr>
 
                                 @foreach($productList as $product)
-                                    <tr>
-                                        <td>{{$product['product_id']}}</td>
-                                        <td>{{$product['product_name']}}</td>
-                                        <td>
-                                            {{$product['status_sell']}}
-                                            <br>
-                                            {{$product['status_purchase']}}
-                                        </td>
-                                        <td>{{$product['warehouse_name']}}</td>
-                                        <td>{{$product['stock']}}</td>
-                                        <td>{{$product['nature']}}</td>
-                                        <td>{{$product['selling_price']}}</td>
-                                        <td>{{$product['product_description']}}</td>
-                                        <td>
-                                        <img src="/upload/Product/{{$product['image']}}" alt="Potato" width="200" height="200"> 
-                                        </td>
-                                        <td>{{$product['product_condition']}}</td>
-                                        <td>
-                                        <a href="/product/edit/{{$product['product_id']}}" class="btn btn-success mb-2">Update</a>
-                                        <a href="/product/delete/{{$product['product_id']}}" class="btn btn-danger">Delete</a>
-                                        </td>
-                                    </tr>
+                                    @if($product['product_condition'] == "Good")
+                                        <tr>
+                                            <td>{{$product['product_id']}}</td>
+                                            <td>{{$product['product_name']}}</td>
+                                            <td>
+                                                {{$product['status_sell']}}
+                                                <br>
+                                                {{$product['status_purchase']}}
+                                            </td>
+                                            <td>{{$product['warehouse_name']}}</td>
+                                            <td>{{$product['stock']}}</td>
+                                            <td>{{$product['nature']}}</td>
+                                            <td>{{$product['selling_price']}}</td>
+                                            <td>{{$product['product_description']}}</td>
+                                            <td>
+                                            <img src="/upload/Product/{{$product['image']}}" alt="Potato" width="200" height="200"> 
+                                            </td>
+                                            <td>{{$product['product_condition']}}</td>
+                                            <td>
+                                            <a href="/product/edit/{{$product['product_id']}}" class="btn btn-success mb-2">Update</a>
+                                            <a href="/product/delete/{{$product['product_id']}}" class="btn btn-danger">Delete</a>
+                                            </td>
+                                        </tr>
+                                    @endif
                                 @endforeach
                             </table>
                         </div>
